+++ shared-components/statviz/components/visualizations/beneficiaries/DemographicDataContainer.tsx
@@ -2,32 +2,11 @@ import { useQuery } from "@apollo/client";
 import { Spinner } from "@chakra-ui/react";
 import { useParams } from "react-router-dom";
 
-import { graphql } from "../../../../../graphql/graphql";
+import { DEMOGRAPHIC_QUERY } from "../../../queries/queries";
 import DemographicFilterContainer from "./DemographicFilterContainer";
 import ErrorCard, { predefinedErrors } from "../../ErrorCard";
 import NoDataCard from "../../NoDataCard";
 import type { BeneficiaryAppliedFilters } from "../../../utils/dashboardFilters";
-
-export const DEMOGRAPHIC_QUERY = graphql(`
-  query BeneficiaryDemographics($baseId: Int!) {
-    beneficiaryDemographics(baseId: $baseId) {
-      facts {
-        count
-        createdOn
-        age
-        gender
-        tagIds
-      }
-      dimensions {
-        tag {
-          id
-          name
-          color
-        }
-      }
-    }
-  }
-`);
 
 interface DemographicDataContainerProps {
   appliedFilters: BeneficiaryAppliedFilters;
