+++ shared-components/statviz/components/visualizations/movedBoxes/MovedBoxes.test.tsx
@@ -1,5 +1,6 @@
 import { it, expect } from "vitest";
-import MovedBoxesDataContainer, { MOVED_BOXES_QUERY } from "./MovedBoxesDataContainer";
+import MovedBoxesDataContainer from "./MovedBoxesDataContainer";
+import { MOVED_BOXES_QUERY } from "../../../queries/queries";
 import { render, screen } from "../../../../tests/testUtils";
 import { GraphQLError } from "graphql";
 import { defaultMovementFilters } from "../../../utils/dashboardFilters";
